Consolidate definition files/Disable "No Customer" feature
As far as I can see the "No Customer" option is not working.. The reason
is pretty simple.. The feature doesn't appear to work, and because my
client doesn't need it, I spent NO TIME trying to run down why because
it would be a WASTE of my time.. So if someone wants to take a run at
fixing this, feel free.. Not sure if I submit this update that I'll be
including the "No Customer" feature. IMHO, anyone who is taking phone or
other manual orders should use one of the manual order modules (by lat9
or lhungil) to properly add an order to their store. THEN use this
module to manage the drop ship notification/management for those orders.
Trying to use this module to do BOTH seems a little counter-productive..